requirements.txt
dacite
matplotlib
pandas
pyyaml
slack_sdk
